backend/pyproject.toml
[project]
name = "xqt5-ai-plattform-backend"
version = "0.1.0"
description = "FastAPI backend for a Coolify-deployed AI hub"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "pydantic>=2.9.0",
  "python-multipart>=0.0.6",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.0.0",
  "slowapi>=0.1.9",
  "supabase>=2.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
